+++ src/robot.h
@@ -104,7 +104,7 @@ void save_sensor(struct sensor *cur_sensor, FILE *fp, int savegame);
 void clear_robot(struct robot *cur_robot);
 void clear_scroll(struct scroll *cur_scroll);
 void clear_sensor(struct sensor *cur_sensor);
-void reallocate_scroll(struct scroll *scroll, int size);
+void reallocate_scroll(struct scroll *scroll, size_t size);
 int find_robot(struct board *src_board, const char *name,
  int *first, int *last);
 void send_robot(struct world *mzx_world, char *name, const char *mesg,
